revert: remove auth end point
