+++ .changeset/cool-points-jam.md
@@ -1,0 +1,5 @@
+---
+"@read-frog/extension": patch
+---
+
+fix: provider test deeplx logic
