+++ .changeset/poor-parks-decide.md
@@ -1,0 +1,5 @@
+---
+"@read-frog/website": patch
+---
+
+revert: remove auth end point
